+++ NEWS.md
@@ -1,3 +1,56 @@
+# EHDEN HMB v0.1.8
+
+* Fix pid and std file names in cohort definitions, changes anticipated build order
+
+# EHDEN HMB v0.1.7
+
+* Correct event in time to procedure intervention
+
+# EHDEN HMB v0.1.6
+
+* Minor bug fix to treatment patterns analysis
+* Correct typo in functions
+
+# EHDEN HMB v0.1.5
+
+* Correct bug in analysis settings (Issue #12)
+* Update package website
+* Add yearly incidence to `Incidence Analysis` (Issue #13)
+* Add R scripts for data exchange and CD preview (Issue #14)
+* update cohort definitions from Siir
+    - copperIUDproc
+    - copperIUDdrug
+    - disorderOfOvary
+
+# EHDEN HMB v0.1.4
+
+* Fix bug in baseline characteristics; typo in analysis settings (Issue #10)
+* Corrections to the study website
+
+
+# EHDEN HMB v0.1.3
+
+* Update cohort definitions
+    - add baseline procedures
+    - add baseline drugs
+* Include procedure prevalence at baseline [-365,0]
+
+# EHDEN HMB v0.1.2
+
+* Update cohort definitions
+    - rerun drug exposure with Capr
+    - rerun procedures with Capr
+    - add time restraints to nsaids
+    - add hmb cohort without hysterectomy censor
+* Add secondary treatment sequence including procedures 
+
+# EHDEN HMB v0.1.1
+
+* Add study website
+* Update cohort definitions for denominator
+* Addition of executeStudy.R file to automate execution
+* add age strata [under 30, 30-45, 45-55]
+
 # EHDEN HMB v0.1.0
 
 * First release of study code for EHDEN HMB